# swagger_2.yaml
...
swagger: '2.0'
info:
  title: BASF AI-SQL Analytics API
  description: |
    API para análisis dinámico de datos de importación BASF con capacidades de IA.
    
    **Nuevas capacidades v4.0:**
    - Ejecución SQL segura generada por IA
    - Análisis de 40+ columnas de datos de importación
    - Validación y sanitización automática de queries
    - Optimizado para Microsoft Copilot Studio
    
    **Endpoints principales:**
    - `/esquema-bd` - Metadata completa para IA
    - `/ejecutar-sql` - Ejecutar SQL dinámico generado por IA
    - `/validar-sql` - Validar consultas antes de ejecutar
  version: '4.0'
host: basf-1.onrender.com
basePath: /
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /ejecutar-sql:
    post:
      summary: Ejecutar SQL generado por IA de Copilot
      operationId: ejecutarSQL
      description: |
        Ejecuta consultas SQL generadas por la IA de Microsoft Copilot Studio.
        Incluye validación de seguridad y optimización automática.
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - instruccion_sql
            properties:
              instruccion_sql:
                type: string
                description: Consulta SQL generada por IA
                example: 'SELECT "PRODUCTO", AVG("FLETE"::numeric) FROM basf_import_data GROUP BY "PRODUCTO" LIMIT 20'
              parametros:
                type: object
                description: Parámetros adicionales para la consulta
              contexto:
                type: string
                description: Contexto de la pregunta original del usuario
                example: "Pregunta: ¿Cuál es el flete promedio por producto?"
      responses:
        '200':
          description: Resultado de la consulta SQL
          schema:
            type: object
            properties:
              status:
                type: string
                enum: [exitoso, sql_invalido, timeout, sin_resultados, error_ejecucion]
              mensaje:
                type: string
              contexto:
                type: string
              resumen:
                type: object
                description: Resumen inteligente de los resultados
              datos:
                type: array
                items:
                  type: object
                description: Resultados de la consulta
              total_registros:
                type: integer
              sql_ejecutado:
                type: string
                description: SQL sanitizado que se ejecutó
              metadata:
                type: object
                properties:
                  columnas:
                    type: array
                    items:
                      type: string
                  tipos_datos:
                    type: object
  /validar-sql:
    get:
      summary: Validar SQL sin ejecutar
      operationId: validarSQL
      description: Útil para debugging y validación previa de consultas SQL
      parameters:
        - name: sql
          in: query
          required: true
          type: string
          description: SQL a validar
        - name: mostrar_explicacion
          in: query
          type: boolean
          default: false
          description: Incluir explicación detallada de la validación
      responses:
        '200':
          description: Resultado de la validación
          schema:
            type: object
            properties:
              sql_original:
                type: string
              es_valido:
                type: boolean
              mensaje_validacion:
                type: string
              sql_sanitizado:
                type: string
...
securityDefinitions: {}
security: []
...